% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geDiffs.R
\name{geDiffs}
\alias{geDiffs}
\title{Gene Expression Differences}
\usage{
geDiffs(gene, v, cpm, n = 2000, cutOff = 1)
}
\arguments{
\item{gene}{Gene commonly mutated in AML}

\item{v}{Variant dataframe v made by mkBeatAML}

\item{cpm}{Counts per million dataframe cpm made by mkBeatAML}

\item{n}{Top n most differentially expressed genes}

\item{cutOff}{Counts per million reads below which a gene is ignored (median across samples)}
}
\value{
Returns Limma's topTable(X, coef=2, adjust="BH",number=n)
}
\description{
Uses Limma to find genes differential expressed when a gene is mutated vs. not (reference).
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") # 672  specimens (rows in clin) 
geDiffs("TET2",v,cpm)
}
\author{
Tom Radivoyevitch
}
