% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geQrtMutCorXL.R
\name{geQrtMutCorXL}
\alias{geQrtMutCorXL}
\title{Excel File of Gene Expression 1st vs 4th quartile correlations with mutations}
\usage{
geQrtMutCorXL(d, v, cpm, f = "~/Results/AML/geQrtMutCorXL.xlsx",
  genes = names(attr(d, "topv")))
}
\arguments{
\item{d}{Clinical dataframe made by muts() after tidyClin() (set number of most mutated genes in muts call}

\item{v}{Variant dataframe v made by mkBeatAML}

\item{cpm}{Counts per million dataframe cpm made by mkBeatAML}

\item{f}{Name of Excel file to write Fisher.test() OR estimates and P values}

\item{genes}{Loops through 1st vs 4th quartiles of expression of genes included in this character vector}
}
\value{
Invisibly, a list of dataframes, one for each sheet in the Excel file.
}
\description{
Loops over gene list to find mutations that correlate with low levels of gene expression. Odd Ratios are postive if
mutations correlate with low levels of expression of the gene.
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") 
d=tidyClin(clin) 
(d=muts(d,v,av,n=14))# crashes with n>14 
geQrtMutCorXL(d,v,cpm,genes=genesE)
}
\author{
Tom Radivoyevitch
}
