% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpr.R
\name{getExpr}
\alias{getExpr}
\title{Get Expression and Mutation Status for Two Genes}
\usage{
getExpr(geneE, geneM, v, cpm)
}
\arguments{
\item{geneE}{Gene Expression of interest}

\item{geneM}{Gene commonly mutated in AML}

\item{v}{Variant dataframe v made by mkBeatAML}

\item{cpm}{Counts per million dataframe cpm made by mkBeatAML}
}
\value{
Returns dataframe with gene expression and mutation status for geneE and geneM
}
\description{
Returns a dataframe with expression values and mutation status for two genes
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") # 672  specimens (rows in clin) 
getExpr("FLT3","FLT3",v,cpm)

}
\author{
Tom Radivoyevitch
}
