% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geVAFcor.R
\name{geVAFcor}
\alias{geVAFcor}
\title{Gene Expression Correlations with VAFs}
\usage{
geVAFcor(gene, v, cpm, N = 2000, cutOff = 1)
}
\arguments{
\item{gene}{Gene commonly mutated in AML}

\item{v}{Variant dataframe v made by mkBeatAML}

\item{cpm}{Counts per million dataframe cpm made by mkBeatAML}

\item{N}{Top N genes most correlated with VAF}

\item{cutOff}{Counts per million reads below which a gene is ignored (median across samples)}
}
\value{
Returns Limma's topTable(X, coef=2, adjust="BH",number=n) where coef=2 is the linear slope parameter
}
\description{
Uses Limma to find genes whosw expression correlates with VAF of a specific mutation.
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") # 672  specimens (rows in clin) 
geVAFcor("TET2",v,cpm)
}
\author{
Tom Radivoyevitch
}
