% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geVAFcorXL.R
\name{geVAFcorXL}
\alias{geVAFcorXL}
\title{Excel File of Gene Expression Correlations with VAFs}
\usage{
geVAFcorXL(d, v, cpm, f = "~/Results/AML/topGEvafCor.xlsx", N = 1000)
}
\arguments{
\item{d}{Clinical dataframe made by muts() after tidyClin() (set number of most mutated genes in muts call}

\item{v}{Variant dataframe v made by mkBeatAML}

\item{cpm}{Counts per million dataframe cpm made by mkBeatAML}

\item{f}{Name of Excel file to write DGE.}

\item{N}{Top N most differentially expressed genes in each sheet}
}
\value{
Returns Limma's topTable(X, coef=2, adjust="BH",number=n)
}
\description{
Loops geVAFcor() over top n mutated genes making sheets of top N Limma-called VAF-correlated genes.
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") 
(d=tidyClin(clin)) #672 rows in clin, one for each measurement; 562 rows in d, one for each patient
(d=muts(d,v,av,n=10))# adds muts and vafs of top n genes and point mut counts from all variants
geVAFcorXL(d,v,cpm,f="~/Results/AML/topGEvafCor.xlsx")#1st sht = clinical, rest = expr vs vafs
}
\author{
Tom Radivoyevitch
}
