% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geXL.R
\name{geXL}
\alias{geXL}
\title{Excel File of Gene Expression Differences}
\usage{
geXL(d, v, cpm, f = "~/Results/AML/topGE.xlsx", N = 1000)
}
\arguments{
\item{d}{Clinical dataframe made by muts() after tidyClin() (set number of most mutated genes in muts call}

\item{v}{Variant dataframe v made by mkBeatAML}

\item{cpm}{Counts per million dataframe cpm made by mkBeatAML}

\item{f}{Name of Excel file to write DGE.}

\item{N}{Top N most differentially expressed genes in each sheet}
}
\value{
Returns Limma's topTable(X, coef=2, adjust="BH",number=n)
}
\description{
Loops geDiffs() over top n mutated genes making sheets of top N Limma-called differentially expressed genes.
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") # 672  specimens (rows in clin) 
geDiffs("TET2",v,cpm)
}
\author{
Tom Radivoyevitch
}
