% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkBeatAML.R
\name{mkBeatAML}
\alias{mkBeatAML}
\title{Make Beat AML R Binary}
\usage{
mkBeatAML(beatHome = "~/data/BeatAML",
  beatFile = "41586_2018_623_MOESM3_ESM.xlsx", outFile = "beatAML")
}
\arguments{
\item{beatHome}{Location of Beat AML excel file.}

\item{beatFile}{Name of Beat AML file. Default is 41586_2018_623_MOESM3_ESM.xlsx.}

\item{outFile}{Name of binary output file, which will be placed in \code{beatHome}.}
}
\value{
Nothing returned. Tibbles v (WES variants), cpm (RNAseq counts per million) and clin are in \code{outFile}.
}
\description{
Converts sheets 5,7, 9, 12 1nd 13 in the Supplementary Tables excel file 
of J.W. Tyner et al. (Nature 562 2018) into tibbles in an R binary file.
}
\note{
The goal here is simply to make reading the files into R faster
}
\examples{
library(AMLbeatR)
mkBeatAML()
}
\author{
Tom Radivoyevitch
}
