% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkListColTib.R
\name{mkListColTib}
\alias{mkListColTib}
\title{Make Beat AML R Binary}
\usage{
mkListColTib(clin = clin, v = v, cpm = cpm)
}
\arguments{
\item{clin}{Clinical tibble made by mkBeatAML}

\item{v}{Variant tibble made by mkBeatAML}

\item{cpm}{Counts per million reads tibble made by mkBeatAML}
}
\value{
Tibble with patients as rows and lab ids nested within a tibbles using list columns
}
\description{
Converts sheets 5,7, 9, 12 1nd 13 in the Supplementary Tables excel file 
of J.W. Tyner et al. (Nature 562 2018) into tibbles in an R binary file.
}
\note{
Inspired by Chapter 25 in R4DS, with countries as patients.
}
\examples{
library(AMLbeatR)
mkBeatAML()
}
\author{
Tom Radivoyevitch
}
