% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCin.R
\name{tidyClin}
\alias{tidyClin}
\title{Tidy up Clinical Data}
\usage{
tidyClin(clin = clin)
}
\arguments{
\item{clin}{Clinical dataframe made by mkBeatAML}
}
\value{
Tibble with one patient per row, sorted by patient ids, with lab ids et al nested in a list column called data
}
\description{
Organizes clin as a tibble sorted by patient ids, with one row per patient and a list column data for multiple measurements
}
\note{
Inspired by Chapter 25 in R4DS, with countries as patients.
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") # 672  specimens (rows in clin) 
(d=tidyClin(clin)) # collected from 562 patients (rows in d)
(D=d\%>\%unnest()) 
(D=d\%>\%unnest(timeNsrc))
(D=d\%>\%unnest(genetics))
(D=d\%>\%unnest(inputs))
(D=d\%>\%unnest(outputs))

}
\author{
Tom Radivoyevitch
}
