% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkBeatAML.R
\name{mkBeatAML}
\alias{mkBeatAML}
\title{Make Beat AML R Binary}
\usage{
mkBeatAML(beatHome = "~/data/BeatAML",
  beatFile = "41586_2018_623_MOESM3_ESM.xlsx", outFile = "beatAML")
}
\arguments{
\item{beatHome}{Location of Beat AML excel file.}

\item{beatFile}{Name of Beat AML file. Default is 41586_2018_623_MOESM3_ESM.xlsx.}

\item{outFile}{Name of binary output file, which will be placed in \code{beatHome}.}
}
\value{
Nothing returned. The dataframes v (WES variants), cpm (RNAseq counts per million) and clin are in \code{outFile}.
}
\description{
Converts sheets 5, 7, and 9 in the Supplementary Tables excel file 
of J.W. Tyner et al. (Nature 562 2018) into dataframes clin, v and cpm in an R binary file.
}
\note{
The goal here is simply to make reading the files into R faster via a binary file. Slapped on 
to clin are DNA and RNA TRUE/FALSE columns via sheets 12 and 13. These are more trustworth than the columns
exomeSeq and rnaSeq in clin
}
\examples{
library(AMLbeatR)
mkBeatAML()
load("~/data/BeatAML/BeatAML.RData")
(vids=sort(unique(v$labId))) #608 have at least one variant
(eids=names(cpm)[-c(1:2)]) #451 RNAseq measurements go 12-xx to 16-xx
(cvids=clin$LabId[clin$DNA]) #622 had DNAseq done on them =>14 AMLs with no variants
(int=intersect(eids,cvids)) #405 samples have RNA and DNA readouts (group of interest)
(int1=intersect(eids,vids)) #399 => 6 with both had no variants 
}
\author{
Tom Radivoyevitch
}
