% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkBeatAML.R
\name{mkBeatAML}
\alias{mkBeatAML}
\title{Make Beat AML R Binary}
\usage{
mkBeatAML(beatHome = "~/data/BeatAML",
  beatFile = "41586_2018_623_MOESM3_ESM.xlsx", outFile = "beatAML")
}
\arguments{
\item{beatHome}{Location of Beat AML excel file.}

\item{beatFile}{Name of Beat AML file. Default is 41586_2018_623_MOESM3_ESM.xlsx.}

\item{outFile}{Name of binary output file, which will be placed in \code{beatHome}.}
}
\value{
Tibble (big data file) saved as d in \code{outFile}.
}
\description{
Converts sheets in the Supplementary Tables excel file 
of J.W. Tyner et al. (Nature 562 2018) into a list-column tibble.
}
\note{
Inspired by Chapter 25 in R4DS, with countries as patients.
}
\examples{
library(AMLbeatR)
mkBeatAML()
}
\author{
Tom Radivoyevitch
}
