% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{muts}
\alias{muts}
\title{Add Mutations to Clinical Data}
\usage{
muts(d, v, n = 10)
}
\arguments{
\item{d}{Clinical dataframe made by tidyClin}

\item{v}{Variant dataframe made by mkBeatAML}

\item{n}{Top n most frequent mutations are encodes in the muts field}
}
\value{
d with additional columns of muts and vafs separated by / and point mutantion counts
}
\description{
Adds mutation information from the variance dataframe v to the genetics list column of clin
}
\examples{
library(AMLbeatR)
load("~/data/BeatAML/BeatAML.RData") # 672  specimens (rows in clin) 
(d=tidyClin(clin)) # collected from 562 patients (rows in d)
(d=muts(d,v)) 
(D=d\%>\%unnest()) 
}
\author{
Tom Radivoyevitch
}
